+++ processing/scale.go
@@ -8,7 +8,12 @@ import (
 
 func scale(pctx *pipelineContext, img *vips.Image, po *options.ProcessingOptions, imgdata *imagedata.ImageData) error {
 	if pctx.wscale != 1 || pctx.hscale != 1 {
-		if err := img.Resize(pctx.wscale, pctx.hscale); err != nil {
+		wscale, hscale := pctx.wscale, pctx.hscale
+		if (pctx.angle+po.Rotate)%180 == 90 {
+			wscale, hscale = hscale, wscale
+		}
+
+		if err := img.Resize(wscale, hscale); err != nil {
 			return err
 		}
 	}